update books list on successful import
